make GetRandom return a float in the range 0 to 1 and rename the GetRandom script function to Random

+++ src/system/system.hpp
@@ -6,7 +6,7 @@
 
 int  GetTime();
 int  GetTicks();
-int  GetRandom();
+float GetRandom();
 void ThreadSleep(int ms);
 
 bool InitSystem(const Log& log);
